splitting up testutil package code to make it easier to avoid circular dependencies
