+++ src/server/pkg/testpachd/mock_pachd.go
@@ -1,4 +1,4 @@
-package testutil
+package testpachd
 
 import (
 	"context"
